Update
This is the final update to my NixOS config, due to Nix being hard to
setup for projects quickly I'm going to switch back to ArchLinux
tomorrow as it's easy to stab yourself in the foot with Arch compared to
Nix it's easier and generally faster to set things up and run smoothly
for projects. Another thing that's really hurt with Nix is my root
partition is 60GB and without having 24GB of it free I occasionally run
out of storage on the root partion when upgrading, that's just silly and
really annoying in addition to that the upgrade process is 15 minutes
long. I reallyed like using Nix for the 4ish months I've used it for,
but Arch is still my preferred and Linux distro I enjoy using the most.
I probably should've stuck with Nix for longer, but I'll be back to try
it again later. Maybe when Flakes are stable and other stuff has
improved.